🚚 move component
