+++ src/components/domain/Page/PageList/PageList.tsx
@@ -4,7 +4,7 @@ import { Grid, Pagination } from '@nextui-org/react';
 import { Page } from '~/domains/Page';
 
 import { NoPageAlert } from '~/components/domain/Page';
-import { PageCard } from '~/components/domain/Page/molecules/PageCard';
+import { PageCard } from '~/components/domain/Page/PageCard';
 
 import { usePagePagination } from '~/hooks/Page';
